+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [7.7.0] - 2026-08-18
+
+_Stable release based on [7.7.0-rc.1]._
+
 ## [7.7.0-rc.1] - 2026-08-18
 
 ### Added
@@ -1292,6 +1296,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[7.7.0]: https://https://github.com/internetguru/laravel-user/compare/v7.6.0...v7.7.0
 [7.7.0-rc.1]: https://github.com/internetguru/laravel-user/releases/tag/v7.6.0
 [7.6.0]: https://https://github.com/internetguru/laravel-user/compare/v7.5.2...v7.6.0
 [7.6.0-rc.1]: https://github.com/internetguru/laravel-user/releases/tag/v7.5.2